+++ src/lib/core/index.ts
@@ -3,6 +3,7 @@ export * as utils from './utils';
 export * from './utility/position';
 export * from './utility/array';
 export * from './utility/bounds';
+export * from './utility/find';
 
 export { v4 as uuid } from 'uuid';